update: move Router.cs to Main/.
