+++ Assets/ConnectApp/Components/CustomDialog.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using ConnectApp.Constants;
+using ConnectApp.Main;
 using RSG;
 using Unity.UIWidgets.animation;
 using Unity.UIWidgets.foundation;
