+++ Assets/ConnectApp/Components/CustomSnackBar.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using ConnectApp.Constants;
+using ConnectApp.Main;
 using Unity.UIWidgets.animation;
 using Unity.UIWidgets.async;
 using Unity.UIWidgets.foundation;
