+++ Assets/ConnectApp/Components/ShareSheet.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using ConnectApp.Constants;
+using ConnectApp.Main;
 using ConnectApp.Plugins;
 using Unity.UIWidgets.foundation;
 using Unity.UIWidgets.gestures;
