+++ Assets/ConnectApp/Redux/Reducers/Reducer.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using ConnectApp.Components;
+using ConnectApp.Main;
 using ConnectApp.Models.Model;
 using ConnectApp.Models.State;
 using ConnectApp.redux.actions;
