+++ Assets/ConnectApp/Screens/MainScreen.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using ConnectApp.Components;
 using ConnectApp.Constants;
+using ConnectApp.Main;
 using ConnectApp.redux;
 using ConnectApp.Utils;
 using Unity.UIWidgets.widgets;
