+++ Assets/ConnectApp/Screens/VideoViewScreen.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using ConnectApp.Components;
 using ConnectApp.Constants;
+using ConnectApp.Main;
 using Unity.UIWidgets.foundation;
 using Unity.UIWidgets.painting;
 using Unity.UIWidgets.rendering;
